% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-aihw-api.R
\docType{data}
\name{base_url}
\alias{base_url}
\title{Base URL for the MyHospitals API}
\format{
An object of class \code{character} of length 1.
}
\usage{
base_url
}
\description{
Base URL for the MyHospitals API
}
\keyword{datasets}
