% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-aihw-api.R
\name{call_aihw_api}
\alias{call_aihw_api}
\title{Call the AIHW API}
\usage{
call_aihw_api(api_url)
}
\arguments{
\item{api_url}{url}
}
\value{
something useful, usually a \code{data.frame}
}
\description{
Call the AIHW API
}
\examples{
\donttest{
call_aihw_api("measure-downloads/myh-adm")
}
}
