% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{download_all_aihw_data}
\alias{download_all_aihw_data}
\title{Download all AIHW MyHospitals data to a single .rds file}
\usage{
download_all_aihw_data(file = "aihw_data.rds", force = FALSE)
}
\arguments{
\item{file}{Path to save data (default: "aihw_data.rds")}

\item{force}{Force redownload, even if file exists and is current (default: FALSE)}
}
\value{
List containing all AIHW data with elements: datasets, measures, reporting_units, etc.
}
\description{
Download all AIHW MyHospitals data to a single .rds file
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Download all data to default file
all_data <- download_all_aihw_data()

# Force fresh download
all_data <- download_all_aihw_data(force = TRUE)
\dontshow{\}) # examplesIf}
}
