% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_all_flat_data}
\alias{get_all_flat_data}
\title{Get all flat data across all measure categories}
\usage{
get_all_flat_data()
}
\value{
data.frame of all flat data
}
\description{
Get all flat data across all measure categories
}
