% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_hospital_mappings}
\alias{get_hospital_mappings}
\title{Get mappings and locations for hospitals.}
\usage{
get_hospital_mappings()
}
\value{
data
}
\description{
Get mappings and locations for hospitals.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_hospital_mappings()
\dontshow{\}) # examplesIf}
}
