% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_measure_categories}
\alias{get_measure_categories}
\title{Get set of measure categories which can be used in \code{read_flat_data_extract()}.}
\usage{
get_measure_categories()
}
\value{
data
}
\description{
Get set of measure categories which can be used in \code{read_flat_data_extract()}.
}
\examples{
\donttest{
get_measure_categories()
}
}
