% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-datasets.R
\name{get_measure_data}
\alias{get_measure_data}
\title{Read datasets by the measure code.}
\usage{
get_measure_data(measure_code)
}
\arguments{
\item{measure_code}{A measure code. For example, "MYH0001".}
}
\value{
data
}
\description{
Read datasets by the measure code.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
get_measure_data("MYH0001")
\dontshow{\}) # examplesIf}
}
