% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_measure_download_codes}
\alias{get_measure_download_codes}
\title{Get set of measure download codes that can be used in \code{get_measure_data()}.}
\usage{
get_measure_download_codes()
}
\value{
data
}
\description{
Get set of measure download codes that can be used in \code{get_measure_data()}.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_measure_download_codes()
\dontshow{\}) # examplesIf}
}
