% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_measure_hierarchy}
\alias{get_measure_hierarchy}
\title{Get the hierarchy of measure categories - measures - reported measures.}
\usage{
get_measure_hierarchy()
}
\value{
data
}
\description{
Get the hierarchy of measure categories - measures - reported measures.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
get_measure_hierarchy()
\dontshow{\}) # examplesIf}
}
