% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-datasets.R
\name{get_measures_from_category}
\alias{get_measures_from_category}
\title{Find the available measures within a measure category.}
\usage{
get_measures_from_category(measure_category_code, trim = TRUE)
}
\arguments{
\item{measure_category_code}{A measure category code. See
\code{get_measure_categories()} for a list.}

\item{trim}{Whether or not to exclude the potentially unnecessary variables.}
}
\value{
data
}
\description{
Find the available measures within a measure category.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_measures_from_category("MYH-CANCER")
\dontshow{\}) # examplesIf}
}
