% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-datasets.R
\name{get_reported_measure_data}
\alias{get_reported_measure_data}
\title{Read datasets by the reported measure code.}
\usage{
get_reported_measure_data(reported_measure_code)
}
\arguments{
\item{reported_measure_code}{A reported measure code. For example, "MYH-RM0037".}
}
\value{
data
}
\description{
Read datasets by the reported measure code.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
get_reported_measure_data("MYH-RM0037")
\dontshow{\}) # examplesIf}
}
