% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_reported_measures}
\alias{get_reported_measures}
\title{Get reported measures.}
\usage{
get_reported_measures()
}
\value{
data
}
\description{
Get reported measures.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
get_reported_measures()
\dontshow{\}) # examplesIf}
}
