% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-datasets.R
\name{get_reporting_unit_data}
\alias{get_reporting_unit_data}
\title{Read datasets by the reporting unit code.}
\usage{
get_reporting_unit_data(reporting_unit_code)
}
\arguments{
\item{reporting_unit_code}{A measure code. For example, "H0012".}
}
\value{
data
}
\description{
Read datasets by the reporting unit code.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) withAutoprint(\{ # examplesIf}
get_reporting_unit_data("H0012")
\dontshow{\}) # examplesIf}
}
