% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-datasets.R
\name{read_dataset_ids}
\alias{read_dataset_ids}
\title{Read (multiple) datasets by id}
\usage{
read_dataset_ids(ids, return_caveats = FALSE)
}
\arguments{
\item{ids}{id of the dataset. See \code{get_datasets()}.}

\item{return_caveats}{A logical. Whether or not to return the caveats for
this data in a separate table. If \code{TRUE}, this will return a list with the
two tables.}
}
\value{
data
}
\description{
Read (multiple) datasets by id
}
\examples{
read_dataset_ids(c(1,2))
}
