% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{read_flat_data_extract}
\alias{read_flat_data_extract}
\title{insert title}
\usage{
read_flat_data_extract(
  measure_category_code,
  measure_code,
  return_caveats = FALSE
)
}
\arguments{
\item{measure_category_code}{test}

\item{measure_code}{test}

\item{return_caveats}{test}
}
\value{
data
}
\description{
insert title
}
\examples{
read_flat_data_extract(measure_category_code = "MYH-CANCER")
}
