% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-myhosp-api.R
\name{call_myhosp_api}
\alias{call_myhosp_api}
\title{Call the MyHospitals API}
\usage{
call_myhosp_api(api_url)
}
\arguments{
\item{api_url}{url}
}
\value{
something useful, usually a \code{data.frame}
}
\description{
Call the MyHospitals API
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
call_myhosp_api("measure-downloads/myh-adm")
\dontshow{\}) # examplesIf}
}
