% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_measure_categories}
\alias{get_measure_categories}
\title{Get set of measure categories which can be used in \code{read_flat_data_extract()}.}
\usage{
get_measure_categories()
}
\value{
data
}
\description{
Get set of measure categories which can be used in \code{read_flat_data_extract()}.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_measure_categories()
\dontshow{\}) # examplesIf}
}
