% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lists.R
\name{get_measure_data}
\alias{get_measure_data}
\title{Get data by measure download code.}
\usage{
get_measure_data(measure_download_code)
}
\arguments{
\item{measure_download_code}{A measure download code. See \code{datasheet_code} in
\code{get_measure_download_codes()}.}
}
\value{
data
}
\description{
Get data by measure download code.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_measure_data("myh-adm")
\dontshow{\}) # examplesIf}
}
