% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-datasets.R
\name{read_dataset_ids}
\alias{read_dataset_ids}
\title{Read (multiple) datasets by id}
\usage{
read_dataset_ids(ids, return_caveats = FALSE, tidy_data = TRUE)
}
\arguments{
\item{ids}{id of the dataset. See \code{get_datasets()}.}

\item{return_caveats}{A logical. Whether or not to return the caveats for
this data in a separate table. If \code{TRUE}, this will return a list with the
two tables.}

\item{tidy_data}{A logical. Whether or not to tidy the data (rename
columns and remove columns which only have one unique value).}
}
\value{
data
}
\description{
Read (multiple) datasets by id
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_dataset_ids(c(1, 2))
\dontshow{\}) # examplesIf}
}
