% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-flat-data-extract.R
\name{read_flat_data_extract}
\alias{read_flat_data_extract}
\title{Get data by measure category.}
\usage{
read_flat_data_extract(
  measure_category_code,
  measure_code,
  return_caveats = FALSE
)
}
\arguments{
\item{measure_category_code}{A measure category code. See
\code{get_measure_categories()} for valid codes.}

\item{measure_code}{A measure code.}

\item{return_caveats}{A logical. Whether or not to return the caveats for
this data in a separate table. If \code{TRUE}, this will return a list with the
two tables.}
}
\value{
data
}
\description{
Get data by measure category.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_flat_data_extract(measure_category_code = "MYH-CANCER")
read_flat_data_extract(measure_category_code = "MYH-CANCER", measure_code = "MYH0001")
\dontshow{\}) # examplesIf}
}
