% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readaihw-package.R
\docType{package}
\name{readaihw-package}
\alias{readaihw}
\alias{readaihw-package}
\title{readaihw: Read Data From the Australian Institute Of Health And Welfare Using The MyHospitals API}
\description{
Provides a wrapper around the MyHospitals API so that it can be accessed directly from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RWParsons/readaihw}
  \item \url{https://rwparsons.github.io/readaihw/}
  \item Report bugs at \url{https://github.com/RWParsons/readaihw/issues}
}

}
\author{
\strong{Maintainer}: Rex Parsons \email{rex.parsons94@gmaqil.com} (\href{https://orcid.org/0000-0002-6053-8174}{ORCID})

}
\keyword{internal}
